Switched the localStorage form save data to the signup page, which makes more sense than having it in the sign in page

+++ authentication-in-express/public/javascripts/clientside.js
@@ -4,14 +4,14 @@ $(document).ready(function() {
     .blur(function(e){
 
       // on blur of any input, save all the data in an object
-      var formData = $('#signin').serializeArray();
+      var formData = $('#signup').serializeArray();
 
       // Set a flag in localStorage to indicate that there is saved data for next page reload
       localStorage.setItem('hasData', true);
 
       $.each(formData, function(i, obj){
         // make sure password is never saved to localStorage
-        if(obj.name !== password){
+        if(obj.name !== 'password' && obj.name !== 'password_confirm'){
           localStorage.setItem(obj.name, obj.value);
         }
 
@@ -28,7 +28,7 @@ $(document).ready(function() {
           "<div class='alert alert-success' role='alert'>Found saved form data in localStorage:</div>"
         );
 
-      var formData = $('#signin').serializeArray();
+      var formData = $('#signup').serializeArray();
 
       $.each(formData, function(i, obj){
         // crafts an attribute selector for each of the previously save form inputs
